+++ README.md
@@ -85,5 +85,5 @@
   - [x] 모든 컨트롤러의 테스트가 성공하면 신규 MVC 프레임워크로 전환한다.
 
 ### 3단계 피드백
-- [ ] ControllerScanner 가 불필요한 인스턴스 필드를 가지지 안도록 변경(유틸성으로 변경하여 정적 메서드로 컨스톨러 인스턴스들을 반환)
+- [x] ControllerScanner 가 불필요한 인스턴스 필드를 가지지 안도록 변경(유틸성으로 변경하여 정적 메서드로 컨스톨러 인스턴스들을 반환)
 - [ ] 테스트 컨트롤러 픽스처의 의도를 표현할 수 있도록 픽스처 패키지로 이동
