+++ src/test/java/core/mvc/tobe/ControllerScannerTest.java
@@ -1,11 +1,11 @@
 package core.mvc.tobe;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import core.annotation.web.Controller;
 import core.mvc.tobe.exception.DuplicatedControllerDefinitionException;
 import java.util.Set;
-import org.assertj.core.api.Assertions;
 import org.assertj.core.api.ThrowableAssert.ThrowingCallable;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -13,16 +13,14 @@ import org.reflections.Reflections;
 
 class ControllerScannerTest {
 
+    private Reflections reflections;
+
     @DisplayName("@Controller 애노테이션이 적용된 클래스를 찾는다.")
     @Test
     void find_controllers_with_controller_annotation() {
-        final Reflections reflections = new Reflections("next.controller");
+        reflections = new Reflections("next.controller");
 
-        final ControllerScanner controllerScanner = new ControllerScanner(reflections);
-
-        controllerScanner.instantiateControllers();
-
-        final Set<Object> controllers = controllerScanner.getControllers();
+        final Set<Object> controllers = ControllerScanner.getControllers(reflections);
 
         final Set<Class<?>> expected = reflections.getTypesAnnotatedWith(Controller.class);
         assertThat(controllers).hasOnlyElementsOfTypes(expected.toArray(new Class<?>[0]));
@@ -31,28 +29,23 @@ class ControllerScannerTest {
     @DisplayName("@Controller 애노테이션이 적용된 클래스를 찾지 못한다.")
     @Test
     void cannot_find_controllers_with_controller_annotation() {
-        final Reflections reflections = new Reflections("next.dao");
+        reflections = new Reflections("next.dao");
 
-        final ControllerScanner controllerScanner = new ControllerScanner(reflections);
+        final Set<Object> controllers = ControllerScanner.getControllers(reflections);
 
-        controllerScanner.instantiateControllers();
-
-        assertThat(controllerScanner.getControllers()).isEmpty();
+        assertThat(controllers).isEmpty();
     }
 
     @DisplayName("@Controller 애노테이션이 적용된 중복된 클래스를 찾는 경우 예외를 발생시킨다")
     @Test
     void cannot_find_duplicated_controllers_with_controller_annotation() {
-        final Reflections reflections = new Reflections("next.duplicatedcontroller");
+        reflections = new Reflections("next.duplicatedcontroller");
 
-        final ControllerScanner controllerScanner = new ControllerScanner(reflections);
+        final ThrowingCallable actual = () -> ControllerScanner.getControllers(reflections);
 
-        final ThrowingCallable actual = controllerScanner::instantiateControllers;
-
-        Assertions.assertThatThrownBy(actual)
+        assertThatThrownBy(actual)
             .isInstanceOf(DuplicatedControllerDefinitionException.class)
             .hasMessage("중복된 컨트롤러 인스턴스가 발견되었습니다: homeController");
 
     }
-
 }
